progression.ai/requirements.txt
pymupdf==1.24.9
pdfminer.six==20231228
